requirements.txt
PyMuPDF==1.22.5
pytesseract==0.3.10
Pillow==10.0.1
pypdf==3.17.4
requests==2.31.0
flask==2.3.3
gunicorn==21.2.0
